requirements.txt
coverage==6.2
dj-database-url==0.5.0
Django==3.1.14
django-cors-headers==3.2.0
django-picklefield==2.0
django-request-logging==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.11.2
fyle==v0.29.0
fyle-rest-auth==1.1.0
gunicorn==20.0.4
isodate==0.6.0
jwt==1.3.1
lxml==4.9.1
psycopg2-binary==2.8.4
pylint==2.7.4
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-mock==3.6.1
python-dateutil==2.8.1
requests==2.25.0
